# Conflicts:
#	lab2/spring_service/src/main/java/com/example/spring_service/configuration/GlobalExceptionHandler.java

+++ lab2/spring_service/src/main/java/com/example/spring_service/model/Coordinates.java
@@ -12,6 +12,6 @@ import java.io.Serializable;
 @Getter
 @Setter
 public class Coordinates implements Serializable {
-    private long x;
-    private Integer y;
+    private Long x;
+    private double y;
 }
